+++ src/components/MarkdownSection.tsx
@@ -1,25 +1,11 @@
-import styled from '@emotion/styled';
+import { Heading } from '@chakra-ui/react';
 import 'github-markdown-css/github-markdown.css';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 
-import { Colors, Fonts } from '@/assets/theme';
-
 interface MarkdownSectionProps {
     source: string;
 }
-
-const Title = styled.h1`
-    font-size: ${Fonts.sizes.h1};
-    font-weight: ${Fonts.weights.heavy};
-    color: ${Colors.content};
-`;
-
-const Subtitle = styled.h2`
-    font-size: ${Fonts.sizes.h2};
-    font-weight: ${Fonts.weights.normal};
-    color: ${Colors.content};
-`;
 
 const MarkdownSection = ({ source }: MarkdownSectionProps): JSX.Element => (
     <div className="markdown-body">
@@ -27,8 +13,8 @@ const MarkdownSection = ({ source }: MarkdownSectionProps): JSX.Element => (
             remarkPlugins={[remarkGfm]}
             children={source}
             components={{
-                h1: ({ ...props }) => <Title {...props} />,
-                h2: ({ ...props }) => <Subtitle {...props} />,
+                h1: () => <Heading as="h1" fontSize="32px" />,
+                h2: () => <Heading fontSize="20px" />,
             }}
         />
     </div>