+++ src/components/layout/applayout/AppLayout.tsx
@@ -1,38 +1,22 @@
-import styled from '@emotion/styled';
+import { Flex } from '@chakra-ui/react';
 
 import { useAppSelector } from '@/hooks';
 
 import Header from '@/components/layout/header/Header';
 
-import { Colors } from '@/assets/theme';
-
 type AppLayoutProps = {
     children: React.ReactNode;
 };
 
-const Container = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    color: ${Colors.content};
-    height: 100vh;
-    width: 100vw;
-`;
-
-const Content = styled.div`
-    flex-grow: 1;
-    overflow-x: auto;
-    overflow-y: auto;
-    display: flex;
-`;
-
 const AppLayout = ({ children }: AppLayoutProps): JSX.Element => {
     const isAuthenticated = useAppSelector((state) => state.user.authenticated);
     return (
-        <Container>
+        <Flex direction="column" align="stretch" height="100vh" width="100vw">
             {isAuthenticated && <Header />}
-            <Content>{children}</Content>
-        </Container>
+            <Flex grow="1" overflowX="auto" overflowY="auto">
+                {children}
+            </Flex>
+        </Flex>
     );
 };